Switch to searchable dropdown
