+++ lib/src/l10n/app_localizations_en.dart
@@ -425,6 +425,13 @@ If this is all OK and you are 18 years old or older, please click \"I agree\". O
   String get locationStepNoCountrySelected => 'None selected';
 
   @override
+  String get locationStepCountrySearchPrompt =>
+      'Please select the country where you are located';
+
+  @override
+  String get locationStepCountryButtonLabel => 'Country:';
+
+  @override
   String get countryCodeAD => 'Andorra';
 
   @override
