+++ lib/src/ui/screens/checkup/steps/temperature.dart
@@ -24,9 +24,12 @@ class _TemperatureStepState extends State<TemperatureStep> {
   static const double _celsiusMin = 21.1;
   static const double _fahrenheitMax = 150.0;
   static const double _fahrenheitMin = 70.0;
-  bool get _isCelsius => _degrees != null && _degrees <= _celsiusMax && _degrees >= _celsiusMin;
+  bool get _isCelsius =>
+      _degrees != null && _degrees <= _celsiusMax && _degrees >= _celsiusMin;
   bool get _isFahrenheit =>
-      _degrees != null && _degrees <= _fahrenheitMax && _degrees >= _fahrenheitMin;
+      _degrees != null &&
+      _degrees <= _fahrenheitMax &&
+      _degrees >= _fahrenheitMin;
   bool get _isValid => _degrees != null && (_isCelsius || _isFahrenheit);
   double _degrees;
 
@@ -103,7 +106,8 @@ class _TemperatureStepState extends State<TemperatureStep> {
           appBar: AppBar(
             title: Text(
               localizations.temperatureStepHowToDialogTitle,
-              style: theme.textTheme.headline.copyWith(color: theme.colorScheme.onBackground),
+              style: theme.textTheme.headline
+                  .copyWith(color: theme.colorScheme.onBackground),
               textAlign: TextAlign.center,
             ),
           ),
@@ -156,7 +160,8 @@ class _TemperatureStepState extends State<TemperatureStep> {
                   child: Center(
                     child: RaisedButton(
                       onPressed: () => Navigator.pop(context),
-                      child: Text(localizations.temperatureStepHowToDialogReturn),
+                      child:
+                          Text(localizations.temperatureStepHowToDialogReturn),
                     ),
                   ),
                 ),
@@ -174,7 +179,8 @@ class _TemperatureStepState extends State<TemperatureStep> {
     return BlocBuilder<CheckupBloc, CheckupState>(
       builder: (context, state) {
         final CheckupStateInProgress checkupState = state;
-        final VitalsResponse existingResponse = checkupState.checkup.vitalsResponses.firstWhere(
+        final VitalsResponse existingResponse =
+            checkupState.checkup.vitalsResponses.firstWhere(
           (VitalsResponse response) => response.id == 'temperature',
           orElse: () => null,
         );
@@ -209,16 +215,19 @@ class _TemperatureStepState extends State<TemperatureStep> {
                             initialValue: existingResponse != null
                                 ? existingResponse.response.toString()
                                 : '',
-                            onChanged: (String value) =>
-                                _updateTemperature(double.parse(value), checkupState),
-                            label: localizations.temperatureStepTemperatureLabel,
+                            onChanged: (String value) => _updateTemperature(
+                                double.parse(value), checkupState),
+                            label:
+                                localizations.temperatureStepTemperatureLabel,
                             suffix: _isValid ? (_isCelsius ? '℃' : '℉') : null,
                             keyboardType: TextInputType.number,
                             inputFormatters: [
-                              WhitelistingTextInputFormatter(RegExp(r'^\d+\.?\d?$')),
+                              WhitelistingTextInputFormatter(
+                                  RegExp(r'^\d+\.?\d?$')),
                             ],
                             autofocus: true,
-                            validator: (String value) => _validateTemperature(value, localizations),
+                            validator: (String value) =>
+                                _validateTemperature(value, localizations),
                           ),
                         ),
                         Spacer(flex: 1),
